+++ lib/backup/backup.dart
@@ -1,1 +1,2 @@
 export 'cubit/backup_cubit.dart';
+export 'view/view.dart';