+++ lib/shared/widgets/widgets.dart
@@ -1,4 +1,4 @@
 export 'expandable_fab.dart';
 export 'filled_card.dart';
 export 'month_paginator.dart';
-export 'update_checker.dart';
+export '../background_worker.dart';